Refactor nutrition service and update ViewModel usage
Replaced NutritionCalculatorService with a new implementation in CalService.swift, updating macronutrient calorie constants and recommendation logic. Updated ViewModel to use the new calorie constants. Removed the old NutritionCalculatorService and renamed DataSeedingService.swift to SeedingService.swift for consistency.
